pyproject.toml
[project]
name = "trucost-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.15.2",
    "authlib>=1.6.0",
    "boto3>=1.38.8",
    "cachetools>=6.1.0",
    "cryptography>=45.0.4",
    "fastapi>=0.115.12",
    "greenlet>=3.2.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.6",
    "pydantic[email]>=2.11.3",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.40",
    "sqlglot>=26.17.1",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"